@@ -1,63 +1,90 @@
-[tool.poetry]
-name = "c7n"
-version = "0.9.44"
-description = "Cloud Custodian - Policy Rules Engine"
-authors = ["Cloud Custodian Project"]
-readme = "README.md"
-homepage = "https://cloudcustodian.io"
-documentation = "https://cloudcustodian.io/docs"
-repository = "https://github.com/cloud-custodian/cloud-custodian"
+[project]
+authors = [
+    {name = "Cloud Custodian Project"},
+]
 license = "Apache-2.0"
-packages = [
-    { include = "c7n" }]
-classifiers=[
-   "License :: OSI Approved :: Apache Software License",
-   "Topic :: System :: Systems Administration",
-   "Topic :: System :: Distributed Computing"]
+requires-python = "<4.0.0,>=3.10.2"
+dependencies = [
+    "argcomplete>=1.12.3",
+    "boto3<2.0.0,>=1.12.31",
+    "cryptography>=44",
+    "jsonschema>=4.18",
+    "python-dateutil<3.0.0,>=2.8.2",
+    "pyyaml>=5.4.0",
+    "tabulate<1.0.0,>=0.9.0",
+    # uv hint: to get latest urllib3
+    "urllib3>2"
+]
+name = "c7n"
+version = "0.9.47"
+description = "Cloud Custodian - Policy Rules Engine"
+readme = "README.md"
+classifiers = [
+    "License :: OSI Approved :: Apache Software License",
+    "Topic :: System :: Systems Administration",
+    "Topic :: System :: Distributed Computing",
+]
 
-[tool.poetry.urls]
+[project.urls]
 "Bug Tracker" = "https://github.com/cloud-custodian/cloud-custodian/issues"
+homepage = "https://cloudcustodian.io"
+repository = "https://github.com/cloud-custodian/cloud-custodian"
+documentation = "https://cloudcustodian.io/docs"
 
-[tool.poetry.scripts]
-custodian = 'c7n.cli:main'
+[project.scripts]
+custodian = "c7n.cli:main"
 
-[tool.poetry.dependencies]
-python = ">=3.9.2,<4.0.0"
-cryptography = ">=44"
-boto3 = "^1.12.31"
-jsonschema = ">=4.18"
-argcomplete = ">=1.12.3"
-python-dateutil = "^2.8.2"
-pyyaml = ">=5.4.0"
-tabulate = "^0.9.0"
-docutils = ">=0.18, <0.19"
-# workaround for: https://github.com/python-poetry/poetry-plugin-export/issues/183
-urllib3 = ">=1.25.4,<1.27"
-# workaround for: https://github.com/pypa/setuptools/issues/4956
-setuptools = "<79"
+[dependency-groups]
+dev = [
+    "click<9.0,>=8.0",
+    "coverage<8,>=7",
+    "docker[websockets]<8.0.0,>=7.1.0",
+    "freezegun<2.0.0,>=1.2.2",
+    "moto<6.0,>=5.0",
+    "openapi-spec-validator<1.0.0,>=0.7.1",
+    "placebo<1,>=0",
+    "pytest<8",
+    "pytest-cov<6,>=3",
+    "pytest-recording<0.14.0,>=0.12.1",
+    "pytest-sugar<1.1.0,>=0.9.2",
+    "pytest-terraform<0.8,>=0.6",
+    "pytest-xdist<4.0,>=3.0",
+    "tomli>=2.2",
+    "twine>=6.0.0",
+    "vcrpy>=4.0.2",
+]
+lint = [
+    "black<25.0, >=23.1",
+    "ruff~=0.11",
+]
+addons = [
+    "aws-xray-sdk<3.0,>=2.14",
+    "jsonpatch<2.0,>=1.25",
+    "psutil>=5.7",
+]
 
-[tool.poetry.group.dev.dependencies]
+[tool.uv.sources]
+c7n = { workspace = true }
 
-ruff = ">=0.3,<0.6"
-docker = {extras = ["websockets"], version = "^7.1.0"}
-black = ">=23.1,<25.0"
-pytest = "<8"
-coverage = "^7"
-placebo = "^0"
-pytest-xdist = "^3.0"
-pytest-cov = ">=3,<6"
-pytest-terraform = ">=0.6,<0.8"
-vcrpy = ">=4.0.2"
-aws-xray-sdk = "^2.5.0"
-jsonpatch = "^1.25"
-psutil = ">=5.7,<7.0"
-twine = ">=3.1.1,<6.0.0"
-pytest-sugar = ">=0.9.2,<1.1.0"
-click = "^8.0"
-freezegun = "^1.2.2"
-pytest-recording = ">=0.12.1,<0.14.0"
-moto = "^5.0"
-openapi-spec-validator = "^0.7.1"
+[tool.uv.workspace]
+members = [
+  "tools/c7n_awscc",
+  "tools/c7n_azure",
+  "tools/c7n_gcp",
+  "tools/c7n_kube",
+  "tools/c7n_left",
+  "tools/c7n_mailer",
+  "tools/c7n_oci",
+  "tools/c7n_openstack",
+  "tools/c7n_org",
+  "tools/c7n_policystream",
+  "tools/c7n_sphinxext",
+  "tools/c7n_tencentcloud",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
 
 [tool.black]
 skip-string-normalization = true
