@@ -1,35 +1,38 @@
-[tool.poetry]
-name = "c7n_awscc"
-version = "0.1.29"
-readme = "readme.md"
-homepage = "https://cloudcustodian.io"
-repository = "https://github.com/cloud-custodian/cloud-custodian"
-description = "Cloud Custodian - AWS Cloud Control Provider"
-documentation = "https://cloudcustodian.io/docs/"
-authors = ["Cloud Custodian Project"]
+[project]
+authors = [
+    {name = "Cloud Custodian Project"},
+]
 license = "Apache-2.0"
-
-classifiers=[
-   "License :: OSI Approved :: Apache Software License",
-   "Topic :: System :: Systems Administration",
-   "Topic :: System :: Distributed Computing"
+requires-python = "<4.0.0,>=3.9.2"
+dependencies = [
+    "boto3",
+    "c7n",
+    "jsonpatch<2.0,>=1.32",
+]
+name = "c7n_awscc"
+version = "0.1.32"
+readme = "readme.md"
+description = "Cloud Custodian - AWS Cloud Control Provider"
+classifiers = [
+    "License :: OSI Approved :: Apache Software License",
+    "Topic :: System :: Systems Administration",
+    "Topic :: System :: Distributed Computing",
 ]
 
-include = ["c7n_awscc/data/*json"]
+[project.urls]
+homepage = "https://cloudcustodian.io"
+repository = "https://github.com/cloud-custodian/cloud-custodian"
+documentation = "https://cloudcustodian.io/docs/"
 
-[tool.poetry.dependencies]
-python = ">=3.9.2,<4.0.0"
-click = "^8.0"
-jsonpatch = "^1.32"
-c7n = {path = "../..", develop = true}
-
-[tool.poetry.build]
-generate-setup-file = false
-script = "build.py"
-
-[tool.poetry.group.dev.dependencies]
-pytest = "<8.0"
+[dependency-groups]
+dev = [
+    "pytest<8.0",
+]
 
 [build-system]
-requires = ["poetry-core>=1.0.0", "boto3"]
-build-backend = "poetry.core.masonry.api"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.hooks.custom]
+path = "build.py"
+dependencies = ["boto3", "requests"]
