@@ -1,35 +1,43 @@
-[tool.poetry]
+[project]
+authors = [
+    {name = "Cloud Custodian Project"},
+]
+license = {text = "Apache-2"}
+requires-python = "<4.0.0,>=3.9.2"
+dependencies = [
+    "c7n",
+    "click>=8.0",
+    "python-hcl2<5.0.0,>=4.3.2",
+    "rich<15.0,>=14.0",
+    "tfparse<1.0,>=0.6",
+]
 name = "c7n_left"
-version = "0.3.30"
+version = "0.3.33"
 description = "Custodian policies for IAAC definitions"
-authors = ["Cloud Custodian Project"]
-license = "Apache-2"
 readme = "README.md"
+classifiers = [
+    "License :: OSI Approved :: Apache Software License",
+    "Topic :: System :: Systems Administration",
+    "Topic :: System :: Distributed Computing",
+]
+
+[project.urls]
 homepage = "https://cloudcustodian.io"
 repository = "https://github.com/cloud-custodian/cloud-custodian"
 documentation = "https://cloudcustodian.io/docs/"
-classifiers = [
-   "License :: OSI Approved :: Apache Software License",
-   "Topic :: System :: Systems Administration",
-   "Topic :: System :: Distributed Computing"
+
+[project.scripts]
+c7n-left = "c7n_left.cli:cli"
+
+[dependency-groups]
+dev = [
+    "pytest<8.0",
+    "pytest-terraform<1.0.0,>=0.7.0",
 ]
 
-[tool.poetry.scripts]
-c7n-left = 'c7n_left.cli:cli'
-
-[tool.poetry.dependencies]
-python = ">=3.9.2,<4.0.0"
-c7n = {path = "../..", develop = true}
-
-click = ">=8.0"
-rich = "^14.0"
-tfparse = "^0.6"
-python-hcl2 = "^4.3.2"
-
-[tool.poetry.group.dev.dependencies]
-pytest = "<8.0"
-pytest-terraform = "^0.7.0"
-
 [build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["c7n_left"]
