@@ -1,26 +1,41 @@
-[tool.poetry]
+[project]
+authors = [
+    {name = "Cloud Custodian Project"},
+]
+license = {text = "Apache-2.0"}
+requires-python = "<4.0.0,>=3.9.2"
+dependencies = [
+    "c7n",
+    "oci<3.0,>=2.106",
+    # uv hint: else it will for some reason downgrade pyopenssl for
+    # reasons not apparent, without an dependency spec precluding
+    # older versions.
+    "pyopenssl>=24"
+]
 name = "c7n_oci"
-version = "0.1.18"
+version = "0.1.21"
 description = "Cloud Custodian - OCI Support"
 readme = "readme.md"
-homepage="https://cloudcustodian.io"
-repository="https://github.com/cloud-custodian/cloud-custodian"
-documentation="https://cloudcustodian.io/docs/"
-authors = ["Cloud Custodian Project"]
-license = "Apache-2.0"
-classifiers=[
-   "License :: OSI Approved :: Apache Software License",
-   "Topic :: System :: Systems Administration",
-   "Topic :: System :: Distributed Computing"
+classifiers = [
+    "License :: OSI Approved :: Apache Software License",
+    "Topic :: System :: Systems Administration",
+    "Topic :: System :: Distributed Computing",
 ]
 
-[tool.poetry.dependencies]
-python = ">=3.9.2,<4.0.0"
-oci = "^2.106"
 
-[tool.poetry.group.dev.dependencies]
-pytest = "<8.0"
+[project.urls]
+homepage = "https://cloudcustodian.io"
+repository = "https://github.com/cloud-custodian/cloud-custodian"
+documentation = "https://cloudcustodian.io/docs/"
+
+[dependency-groups]
+dev = [
+    "pytest<8.0",
+]
 
 [build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["c7n_oci"]
