@@ -392,8 +392,8 @@ class LambdaMode(ServerlessExecutionMode):
             # Lambda passthrough config
             'layers': {'type': 'array', 'items': {'type': 'string'}},
             'concurrency': {'type': 'integer'},
-            'runtime': {'enum': ['python3.8', 'python3.9', 'python3.10',
-                                 'python3.11', 'python3.12']},
+            'runtime': {'enum': ['python3.10', 'python3.11', 'python3.12',
+                                 'python3.13']},
             'role': {'type': 'string'},
             'handler': {'type': 'string'},
             'pattern': {'type': 'object', 'minProperties': 1},
