@@ -50,6 +50,7 @@ ResourceMap = {
   "aws.cloudsearch": "c7n.resources.cloudsearch.CloudSearch",
   "aws.cloudtrail": "c7n.resources.cloudtrail.CloudTrail",
   "aws.cloudwatch-dashboard": "c7n.resources.cw.CloudWatchDashboard",
+  "aws.cloudwatch-synthetics": "c7n.resources.cw.SyntheticsCanary",
   "aws.codebuild": "c7n.resources.code.CodeBuildProject",
   "aws.codebuild-credential": "c7n.resources.code.CodeBuildSourceCredentials",
   "aws.codecommit": "c7n.resources.code.CodeRepository",
@@ -59,10 +60,28 @@ ResourceMap = {
   "aws.codedeploy-group": "c7n.resources.code.CodeDeployDeploymentGroup",
   "aws.codepipeline": "c7n.resources.code.CodeDeployPipeline",
   "aws.composite-alarm": "c7n.resources.cw.CompositeAlarm",
-  "aws.comprehend-document-classifier": "c7n.resources.comprehend.ComprehendDocumentClassifier",
+  "aws.comprehend-document-classification-job":
+    "c7n.resources.comprehend.ComprehendDocumentClassificationJob",
+  "aws.comprehend-document-classifier":
+    "c7n.resources.comprehend.ComprehendDocumentClassifier",
+  "aws.comprehend-dominant-language-detection-job":
+    "c7n.resources.comprehend.ComprehendDominantLanguageDetectionJob",
   "aws.comprehend-endpoint": "c7n.resources.comprehend.ComprehendEndpoint",
+  "aws.comprehend-entities-detection-job":
+   "c7n.resources.comprehend.ComprehendEntitiesDetectionJob",
   "aws.comprehend-entity-recognizer": "c7n.resources.comprehend.ComprehendEntityRecognizer",
+  "aws.comprehend-events-detection-job": "c7n.resources.comprehend.ComprehendEventsDetectionJob",
   "aws.comprehend-flywheel": "c7n.resources.comprehend.ComprehendFlywheel",
+  "aws.comprehend-key-phrases-detection-job":
+  "c7n.resources.comprehend.ComprehendKeyPhrasesDetectionJob",
+  "aws.comprehend-pii-entities-detection-job":
+    "c7n.resources.comprehend.ComprehendPiiEntitiesDetectionJob",
+  "aws.comprehend-sentiment-detection-job":
+    "c7n.resources.comprehend.ComprehendSentimentDetectionJob",
+  "aws.comprehend-targeted-sentiment-detection-job":
+    "c7n.resources.comprehend.ComprehendTargetedSentimentDetectionJob",
+  "aws.comprehend-topics-detection-job":
+    "c7n.resources.comprehend.ComprehendTopicsDetectionJob",
   "aws.config-recorder": "c7n.resources.config.ConfigRecorder",
   "aws.config-rule": "c7n.resources.config.ConfigRule",
   "aws.connect-campaign": "c7n.resources.connect.ConnectCampaign",
@@ -106,6 +125,7 @@ ResourceMap = {
   "aws.eks-nodegroup": "c7n.resources.eks.NodeGroup",
   "aws.elastic-ip": "c7n.resources.vpc.NetworkAddress",
   "aws.elasticache-group": "c7n.resources.elasticache.ElastiCacheReplicationGroup",
+  "aws.elasticache-reserved": "c7n.resources.elasticache.ReservedCacheNodes",
   "aws.elasticache-user": "c7n.resources.elasticache.ElastiCacheUser",
   "aws.elasticbeanstalk": "c7n.resources.elasticbeanstalk.ElasticBeanstalk",
   "aws.elasticbeanstalk-environment": "c7n.resources.elasticbeanstalk.ElasticBeanstalkEnvironment",
@@ -116,9 +136,10 @@ ResourceMap = {
   "aws.emr-security-configuration": "c7n.resources.emr.EMRSecurityConfiguration",
   "aws.emr-serverless-app": "c7n.resources.emr.EMRServerless",
   "aws.eni": "c7n.resources.vpc.NetworkInterface",
-  "aws.event-bus": "c7n.resources.cw.EventBus",
-  "aws.event-rule": "c7n.resources.cw.EventRule",
-  "aws.event-rule-target": "c7n.resources.cw.EventRuleTarget",
+  "aws.event-bus": "c7n.resources.eventbridge.EventBus",
+  "aws.event-rule": "c7n.resources.eventbridge.EventRule",
+  "aws.event-rule-target": "c7n.resources.eventbridge.EventRuleTarget",
+  "aws.eventbridge-pipes": "c7n.resources.eventbridge.EventBridgePipes",
   "aws.firehose": "c7n.resources.kinesis.DeliveryStream",
   "aws.firewall": "c7n.resources.firewall.NetworkFirewall",
   "aws.fis-experiment": "c7n.resources.fis.Experiment",
@@ -168,6 +189,8 @@ ResourceMap = {
   "aws.kafka-config": "c7n.resources.kafka.KafkaClusterConfiguration",
   "aws.kendra": "c7n.resources.kendra.KendraIndex",
   "aws.key-pair": "c7n.resources.vpc.KeyPair",
+  "aws.keyspace": "c7n.resources.keyspaces.Keyspace",
+  "aws.keyspace-table": "c7n.resources.keyspaces.Table",
   "aws.kinesis": "c7n.resources.kinesis.KinesisStream",
   "aws.kinesis-analytics": "c7n.resources.kinesis.AnalyticsApp",
   "aws.kinesis-analyticsv2": "c7n.resources.kinesis.KinesisAnalyticsAppV2",
@@ -176,6 +199,7 @@ ResourceMap = {
   "aws.kms-key": "c7n.resources.kms.Key",
   "aws.lambda": "c7n.resources.awslambda.AWSLambda",
   "aws.lambda-layer": "c7n.resources.awslambda.LambdaLayerVersion",
+  "aws.lambda-event-source-mapping": "c7n.resources.awslambda.LambdaEventSourceMapping",
   "aws.launch-config": "c7n.resources.asg.LaunchConfig",
   "aws.launch-template-version": "c7n.resources.ec2.LaunchTemplate",
   "aws.lex-bot": "c7n.resources.lex.LexBot",
@@ -220,6 +244,8 @@ ResourceMap = {
   "aws.prefix-list": "c7n.resources.vpc.PrefixList",
   "aws.qldb": "c7n.resources.qldb.QLDB",
   "aws.quicksight-account": "c7n.resources.quicksight.QuicksightAccount",
+  "aws.quicksight-dashboard": "c7n.resources.quicksight.QuicksightDashboard",
+  "aws.quicksight-datasource": "c7n.resources.quicksight.QuicksightDataSource",
   "aws.quicksight-group": "c7n.resources.quicksight.QuicksightGroup",
   "aws.quicksight-user": "c7n.resources.quicksight.QuicksightUser",
   "aws.r53domain": "c7n.resources.route53.Route53Domain",
@@ -324,6 +350,8 @@ ResourceMap = {
   "aws.user-pool": "c7n.resources.cognito.CognitoUserPool",
   "aws.vpc": "c7n.resources.vpc.Vpc",
   "aws.vpc-endpoint": "c7n.resources.vpc.VpcEndpoint",
+  "aws.vpc-endpoint-service-configuration": "c7n.resources.vpc.VPCEndpointServiceConfiguration",
+  "aws.client-vpn-endpoint": "c7n.resources.vpc.ClientVpnEndpoint",
   "aws.vpn-connection": "c7n.resources.vpc.VPNConnection",
   "aws.vpn-gateway": "c7n.resources.vpc.VPNGateway",
   "aws.waf": "c7n.resources.waf.WAF",
