@@ -1,6 +1,7 @@
 # Copyright The Cloud Custodian Authors.
 # SPDX-License-Identifier: Apache-2.0
 
+import datetime
 import logging
 
 from azure.keyvault.keys import KeyProperties
@@ -9,6 +10,7 @@ from c7n.filters import Filter
 from c7n.utils import type_schema
 
 from c7n_azure import constants
+from c7n_azure.actions.base import AzureBaseAction
 from c7n_azure.provider import resources
 from c7n_azure.query import ChildResourceManager, ChildTypeInfo
 from c7n_azure.utils import ThreadHelper, ResourceIdParser
@@ -177,8 +179,10 @@ class KeyVaultKeyRotationFilter(Filter):
                 - type: rotation-policy
                   state: Disabled
     """
-    schema = type_schema('rotation-policy', required=['state'],
-                state={'type': 'string', 'enum': ['Enabled', 'Disabled']})
+    schema = type_schema(
+        'rotation-policy', required=['state'],
+        state={'type': 'string', 'enum': ['Enabled', 'Disabled']}
+    )
 
     def process(self, resources, event=None):
         matched = []
@@ -187,6 +191,76 @@ class KeyVaultKeyRotationFilter(Filter):
             client = self.manager.get_client(vault_url=id.vault_url)
             rotation = client.get_key_rotation_policy(id.name)
             if (self.data.get('state') == 'Disabled' and not rotation.id) or \
-               (self.data.get('state')  == 'Enabled' and rotation.id):
+               (self.data.get('state') == 'Enabled' and rotation.id):
                 matched.append(key)
         return matched
+
+
+@KeyVaultKeys.action_registry.register("update")
+class KeyVaultKeyUpdateAction(AzureBaseAction):
+    """
+    Update a key's properties. This does not update its cryptographic material.
+
+    expires_on and not_before require isoformatted dates
+
+    :example:
+
+    .. code-block: yaml
+        policies:
+            - name: update-expires-on
+              resource: azure.keyvault-keys
+              filters:
+                - name: foo
+              actions:
+                - type: update
+                  expires_on: "2025-12-12 00:00:00"
+    """
+
+    schema = type_schema(
+        "update",
+        key_operations={
+            "type": "array",
+            "items": {
+                "type": "string",
+                "enum": [
+                    "decrypt",
+                    "encrypt",
+                    "export",
+                    "import_key",
+                    "sign",
+                    "unwrap_key",
+                    "verify",
+                    "wrap_key"
+                ]
+            },
+        },
+        enabled={"type": "boolean"},
+        tags={"type": "object"},
+        not_before={"type": "string", "format": "date-time"},
+        expires_on={"type": "string", "format": "date-time"},
+    )
+
+    def _process_resource(self, resource):
+        name = resource['vault_id']['resource_id']['name']
+
+        id = KeyProperties(key_id=resource['id'])
+
+        client = self.manager.get_client(vault_url=id.vault_url)
+
+        not_before = self.data.get("not_before")
+        expires_on = self.data.get("expires_on")
+
+        if not_before:
+            not_before = datetime.datetime.fromisoformat(not_before)
+
+        if expires_on:
+            expires_on = datetime.datetime.fromisoformat(expires_on)
+
+        client.update_key_properties(
+            name=name,
+            key_operations=self.data.get("key_operations"),
+            enabled=self.data.get("enabled"),
+            tags=self.data.get("tags"),
+            not_before=not_before,
+            expires_on=expires_on,
+        )
